podcast_summarizer/requirements.txt

streamlit
crewai
crewai-tools
opik
weaviate-client
python-dotenv
openai
yt-dlp
whisper
ffmpeg-python
